fixed integration test
